Move `URLProtocolStub ` to separate file
